# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LARAVEL Rest API DOCUMENTATION",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Author"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/conversation/add": {
            "post": {
                "tags": [
                    "Conversation"
                ],
                "summary": "add new conversation",
                "operationId": "addNewConversation",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "creator",
                                    "member"
                                ],
                                "properties": {
                                    "creator": {
                                        "type": "integer"
                                    },
                                    "member": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/conversation/leave": {
            "post": {
                "tags": [
                    "Conversation"
                ],
                "summary": "leave conversation",
                "operationId": "leaveConversation",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "conId",
                                    "uid"
                                ],
                                "properties": {
                                    "conId": {
                                        "type": "integer"
                                    },
                                    "uid": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/friend/people": {
            "get": {
                "tags": [
                    "Friend"
                ],
                "summary": "search friend user added",
                "operationId": "findPeople",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/friend/un-friend": {
            "post": {
                "tags": [
                    "Friend"
                ],
                "summary": "un friend ",
                "operationId": "unFriend",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "userIdWant",
                                    "userIdBe"
                                ],
                                "properties": {
                                    "userIdWant": {
                                        "type": "integer"
                                    },
                                    "userIdBe": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/friend/add": {
            "post": {
                "tags": [
                    "Friend"
                ],
                "summary": "add friend",
                "operationId": "addFriend",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "userWantAdd",
                                    "userBeAdded"
                                ],
                                "properties": {
                                    "userWantAdd": {
                                        "type": "integer"
                                    },
                                    "userBeAdded": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/messages": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "get message with limit",
                "operationId": "getMessages",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "uid",
                                    "conId",
                                    "offset",
                                    "limit"
                                ],
                                "properties": {
                                    "uid": {
                                        "type": "integer"
                                    },
                                    "conId": {
                                        "type": "integer"
                                    },
                                    "offset": {
                                        "type": "integer"
                                    },
                                    "limit": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/messages/edit": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "edit messages",
                "operationId": "edit",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "uid",
                                    "conId",
                                    "mesId",
                                    "content"
                                ],
                                "properties": {
                                    "uid": {
                                        "type": "integer"
                                    },
                                    "conId": {
                                        "type": "integer"
                                    },
                                    "mesId": {
                                        "type": "integer"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/messages/delete": {
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "delete messages",
                "operationId": "delete",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "uid",
                                    "conId",
                                    "mesId"
                                ],
                                "properties": {
                                    "uid": {
                                        "type": "integer"
                                    },
                                    "conId": {
                                        "type": "integer"
                                    },
                                    "mesId": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "api/account/register": {
            "post": {
                "tags": [
                    "Author"
                ],
                "summary": "Register",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "fullName",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}